pyproject.toml
[tool.poetry]
name = "yadirect-bigquery-connector"
version = "0.1.0"
description = ""
authors = ["tdcau-py <borisov.p.92@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
google-cloud-bigquery = "^3.14.1"
pandas = "^2.1.4"
pyarrow = "^14.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
